Create module for medcatlite and added copywrite notice
