+++ healthchain/pipeline/models/medcatlite/__init__.py
@@ -1,0 +1,3 @@
+from .medcatlite import MedCATLite
+
+__all__ = ["MedCATLite"]
